Server time zone
Fixed the issue where the SendEmails
function uses .toLocaleString() to
retrieve the current date when fetching
tasks. This is because the server has
two time zones: one international and one
national. If this function is not used, the
international time zone is used, and
starting at 10 p.m. in Argentina, the
server uses a different day using
international time

Resolve: R#029